refactor(modoc): added test refactor error and some other files
