+++ crates/cloud-wallet-openid4vc/src/formats/mdoc/parser.rs
@@ -104,11 +104,13 @@ impl ParsedMdoc {
         let name_spaces_val = take_entry(&mut issuer_signed_map, "nameSpaces")?;
         let issuer_auth_val = take_entry(&mut issuer_signed_map, "issuerAuth")?;
 
-        // RFC 9052 §4.2: COSE_Sign1 is encoded as CBOR tag 18.
-        // issuerAuth MUST be #6.18(COSE_Sign1); reject untagged arrays for
-        // strict ISO/COSE compliance.
+        // ISO 18013-5 §9.1.2 CDDL defines issuerAuth as COSE_Sign1 without mandating
+        // the optional CBOR tag 18 (RFC 9052 §4.2). Real-world implementations —
+        // including the official ISO Annex D test vectors — use the untagged form.
+        // Accept both; reject anything that is neither a tag-18 wrapper nor a bare array.
         let cose_payload_val = match issuer_auth_val {
             Value::Tag(18, inner) => *inner,
+            val @ Value::Array(_) => val,
             _ => {
                 return Err(MdocError::UnexpectedCborType {
                     field: "issuerAuth",
