+++ crates/cloud-wallet-openid4vc/test_data/mdoc/issuer_signed.b64url
@@ -1,0 +1,1 @@
+ompuYW1lU3BhY2VzoXFvcmcuaXNvLjE4MDEzLjUuMYbYGFhjpGhkaWdlc3RJRABmcmFuZG9tWCCHmGRbIOogDhn_q6ySYkvuauxjrO7ez7G4AHfSK_wg6XFlbGVtZW50SWRlbnRpZmllcmtmYW1pbHlfbmFtZWxlbGVtZW50VmFsdWVjRG9l2BhYbKRoZGlnZXN0SUQDZnJhbmRvbVggsj9ifomZxwbfDApO2YrXSvmIr2GbS7B4uJBYVT9EYV1xZWxlbWVudElkZW50aWZpZXJqaXNzdWVfZGF0ZWxlbGVtZW50VmFsdWXZA-xqMjAxOS0xMC0yMNgYWG2kaGRpZ2VzdElEBGZyYW5kb21YIMf_owfl3pIeZ7pYeAlHh-iAesjntbOTLSzoDwDz6auvcWVsZW1lbnRJZGVudGlmaWVya2V4cGlyeV9kYXRlbGVsZW1lbnRWYWx1ZdkD7GoyMDI0LTEwLTIw2BhYbaRoZGlnZXN0SUQHZnJhbmRvbVggJgUqQuWIBVeoBsFFmvP7frUF03gVZjKdC2BLhFtfnmhxZWxlbWVudElkZW50aWZpZXJvZG9jdW1lbnRfbnVtYmVybGVsZW1lbnRWYWx1ZWkxMjM0NTY3ODnYGFkEcaRoZGlnZXN0SUQIZnJhbmRvbVgg0JTa12Si653rUhDp2JlkPvvR0GnMMR0ylVFsoLAkQS1xZWxlbWVudElkZW50aWZpZXJocG9ydHJhaXRsZWxlbWVudFZhbHVlWQQS_9j_4AAQSkZJRgABAQEAkACQAAD_2wBDABMNDhEODBMRDxEVFBMXHTAfHRoaHToqLCMwRT1JR0Q9Q0FMVm1dTFFoUkFDX4JgaHF1e3x7SlyGkIV3j214e3b_2wBDARQVFR0ZHTgfHzh2T0NPdnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnb_wAARCAAYAGQDASIAAhEBAxEB_8QAGwAAAwEAAwEAAAAAAAAAAAAAAAUGBAECAwf_xAAyEAABAwMDAgUCAwkAAAAAAAABAgMEAAURBhIhEzEUFVFhcSJBB4GhFjVCUnORssHx_8QAFQEBAQAAAAAAAAAAAAAAAAAAAAH_xAAaEQEBAQADAQAAAAAAAAAAAAAAAUERITFh_9oADAMBAAIRAxEAPwClu94i2iMpx9aSvH0NA_Us-w_3Xnp-8-dwlyOh0NrhRt37s8A5zgetK9R6fjLbuN0dUtbvSyhPZKSABn37Ufh_-5X_AOuf8U0hXeZq8InORLfb3py2iQooOO3fGAePet1i1BHvTbmxCmXWuVoUc4HqDUlbkzJ1_mu6dcEUEEqLpBBBPpg9_wBPWvXTS0tM3mMtC_H9FZK92RxkEfOTTC-mr2tUl10Qbc9KZa5W6FYAHrwDx84p3Z7vHvEPxEfcnadq0q7pNTehun5PcN2O_wBXxt_7XhoZhUqDdY5UUodQlG7GcEhQzQN7zrCLbX0sx20zF_x7XMBPtnByacXG4MW2CuVJJCEjsOST9gKgdVWeNZlw2Y24lSVFa1HJUcivoT6o6Y48WWg2eD1cY_WmGpn9tykIddtL6IqzhLu7v8cYP96qYz6JUdt9o5bcSFJPsai9YRpaoqJDLzCrQgp6bTJAxxjPAx-p70ya1VAgWqApUd9KHWyEIbAVt2nbjJIpg36ivosbDTnQ66nFFITv24wO_Y0lja88RJaZ8u29RYTnr5xk4_lrm-so1KxAkx5keMjnaiSoJUSVAdhn0rHc3rrpm5x1KuTs1t3koXnBweRgk4-RSe9lXlFcA5GaKJyz3KJ4-3vxd_T6qCndjOPyrJp-zeSQlx-v19zhXu2bccAYxk-lFFFLJOjk-MXJt1wegledwSM9_sCCOPat1i05GswcUlannnBtUtQxx6AUUUC5_RSes6YNxeiMu8LaCSQR6dxx85p3ZrRHs0ToR9ysnctau6jRRQYdQ6b88eZc8V0OkCMdPdnP5imVxtzFyhKiyQShX3HdJ9RRRT4J0aIUUJYcuz6oqVZDO3gfHOM9_tVPDitQorcdhO1tsYAoooF190_GvaEFxSmnkcJcTzx6EfcVhiaPSma3JuM96epvG1Kxgcdgck5HtRRSClooooP_2dgYWP-kaGRpZ2VzdElECWZyYW5kb21YIEWZ-BvqorIL0P_MmqA6b5hb76s_a-r_pB5jVM2yqyzkcWVsZW1lbnRJZGVudGlmaWVycmRyaXZpbmdfcHJpdmlsZWdlc2xlbGVtZW50VmFsdWWCo3V2ZWhpY2xlX2NhdGVnb3J5X2NvZGVhQWppc3N1ZV9kYXRl2QPsajIwMTgtMDgtMDlrZXhwaXJ5X2RhdGXZA-xqMjAyNC0xMC0yMKN1dmVoaWNsZV9jYXRlZ29yeV9jb2RlYUJqaXNzdWVfZGF0ZdkD7GoyMDE3LTAyLTIza2V4cGlyeV9kYXRl2QPsajIwMjQtMTAtMjBqaXNzdWVyQXV0aIRDoQEmoRghWQHzMIIB7zCCAZWgAwIBAgIUPEQW7teE87QT5I9W8HWr-m2H64QwCgYIKoZIzj0EAwIwIzEUMBIGA1UEAwwLdXRvcGlhIGlhY2ExCzAJBgNVBAYTAlVTMB4XDTIwMTAwMTAwMDAwMFoXDTIxMTAwMTAwMDAwMFowITESMBAGA1UEAwwJdXRvcGlhIGRzMQswCQYDVQQGEwJVUzBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABKznq3NA5dlkjFpyqab1Z0XHqtQ2oDpD7-p3tfp7iPAZfVfYmD4bN9OlOfTViDZeOMu_W5TWjFR7W8hzHc0vFGujgagwgaUwHgYDVR0SBBcwFYETZXhhbXBsZUBleGFtcGxlLmNvbTAcBgNVHR8EFTATMBGgD6ANggtleGFtcGxlLmNvbTAdBgNVHQ4EFgQUFOKQF6bDViH_x6aGt7ctsGzRI1EwHwYDVR0jBBgwFoAUVPojg6BMKODZMHkiYcgMSIHSwAswDgYDVR0PAQH_BAQDAgeAMBUGA1UdJQEB_wQLMAkGByiBjF0FAQIwCgYIKoZIzj0EAwIDSAAwRQIhAJdxerkBZ0DI17zapJSmLAU7vezOE4PBrKcq0I28BMuyAiA7rYWcE6Y8bRrWfYFNQ-JCXK-Q1CJCLASo7gMEwNOmjVkDotgYWQOdpmd2ZXJzaW9uYzEuMG9kaWdlc3RBbGdvcml0aG1nU0hBLTI1Nmx2YWx1ZURpZ2VzdHOicW9yZy5pc28uMTgwMTMuNS4xrQBYIHUWczO0e2wr-4bszB9DjPV68FU3GsVeHjWeIPJUrc6_AVggZ-U51hOevRMa70QbRFZF3YMbKzdbOQyl72J5sgXtRXECWCAzlDct23gFPzbV2Gl4DmHtoxPUSjkgkq2OBSei-_5VrgNYIC41rTxOUUu2exqdtRznTky5txRuQaxS2snOhrhhPbVVBFgg6lwzBLt8So3LUcTBO2UmT4RVQTQTQgk8ynhuBY-sLVkFWCD65If2i3oOh6dJd05W6eHcOo7Ht35JDSHw4dNHVmGqHQZYIH2D5Qeud9uBXeTYA7iFVdBRHYlMiXQ59XdAVkFqHHUzB1gg8FSaFF8c91y-7_qIHUhX3UONYnzzIXSxcxxMOOEsqTYIWCC2jIr8sqr3xYFBHSh33vFVvi6xIaQrybpbcxI3fgaPZglYIAs1h9HdDCoHo1v7Eg2ZoKv7XfVoZbt_oVzItWpm324MClggyYoXDPNuEau3JOmKdaU0PforbtPfLs-7jvLuVd1ByIELWCC1fdA2eC97FMajD6qq5szVBUzoi9-lGgFrp17aHt6pSAxYIGUfhzaxhID-JSoDIk6gh7XRDKVIUUbGfHSsTsMRLUw6dG9yZy5pc28uMTgwMTMuNS4xLlVTpABYINgLg9JRc8SExWQGEP8aMclJwdk0v0z38Y1SI7Fd1PIcAVggTYDh4uT7JG2XiVQnznAAu1m7JMjNAD7PlL81u9KRfjQCWCCLMx87aFvKNy6FNRolyUhKt6_N8NIjMQVRH3eNmML1RANYIMNDrxvRaQcVQ5Fhq6c3AsR0q_mSsgyftVw2ozbr4BqHbWRldmljZUtleUluZm-haWRldmljZUtleaQBAiABIVggljE9bGPiTjNydCv9saM7osiX3NaKuMdT5PvUjcprf5oiWCAfsyae3UGIV94bOaTkpEuS-khMqnIsIoKI8B0MA6LD1mdkb2NUeXBldW9yZy5pc28uMTgwMTMuNS4xLm1ETGx2YWxpZGl0eUluZm-jZnNpZ25lZMB0MjAyMC0xMC0wMVQxMzozMDowMlppdmFsaWRGcm9twHQyMDIwLTEwLTAxVDEzOjMwOjAyWmp2YWxpZFVudGlswHQyMDIxLTEwLTAxVDEzOjMwOjAyWlhAWeZCBd8eL3CN1tsIR67Xn8fAIB2A-lW63K8uG89ZAuHlpi5IMgRLiQrYWqU_EpE0d11zN1TXy3pBN2au_xPLLg
